requirements.txt
pandas
numpy
requests
scikit-learn
vnstock
python-dotenv